requirements.txt
pytest==7.4.3
hypothesis==6.92.1
boto3==1.34.0
trino==0.23.0
requests==2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyarrow>=14.0.0
minio>=7.2.0
